drop /parse/multi endpoint
Replaced by /parse/multiPart
